Added 'UseDict' method handler
When the requested method name is not found the server will check to see if there is a 'UseDict' listener registered.  This allows users to use other methods to serve up the RPC requested methods than just registering a listener as well as handle methods not found differently than just emitting a 404 error code when the receiving end does not handle that error code well.