Added useDict event listener type to the file
Added a new event listener to the lib/server.js called useDict.  This allows for the consumer to define a custom object of functions that are just plain functions versus being required to register a callback event listener for each method that is being exposed.  This allows for easier testing of the methods as well as the ability to more elegantly handle errors on systems that do not handle 404 error codes well